+++ CHANGES.md
@@ -5,6 +5,8 @@ CHANGELOG
 0.0.3.dev0    (XXXX-XX-XX)
 --------------------------
 
+* Add real_url property to model MapBAseLayer, to get real url for mapbox styles
+
 
 0.0.2         (2019-11-29)
 --------------------------